fix iterator usage
Some sets invalidate iterators on erase(). That means it's not allowed to do this:

sides.erase(it++);

since erase might invalidate all iterators, it's not allowed to use it afterwards. So the code in USE_BTREE

it = sides.erase(it++);

is more correct, but the ++ is not necessary because it will be overwritten anyways. The version that works with any set, no matter if iterators are invalidated or not is this:

it = sides.erase(it)

Only abseil does it differently: https://github.com/abseil/abseil-cpp/blob/20230125.0/absl/container/flat_hash_map.h#L228-L233